缓存层开始增加对有序集合zorder set的支持

+++ commonredis/src/main/java/ywh/common/redis/KeysRao.java
@@ -8,6 +8,6 @@ public interface KeysRao {
 
     public Iterable<String> getKeys(final String pattern);
 
-    public Long keysCount(final String pattern);
+    public Integer keysCount(final String pattern);
 
 }
